Refactor/ useSections hook into store and more components for better maintainance
